[opan] Refactoring feature specs
