+++ spec/features/barito_app_management/create_barito_app_spec.rb
@@ -1,0 +1,64 @@
+require 'rails_helper'
+
+RSpec.feature 'Barito App Management', type: :feature do
+  let(:user_a) { create(:user) }
+  let(:user_b) { create(:user) }
+  let(:admin) { create(:user, :admin) }
+
+  describe 'Create new barito app' do
+    before(:each) do
+      @app_group = create(:app_group, created_by: user_a)
+      create(:infrastructure, app_group: @app_group)
+    end
+
+    context 'As Owner/ As Superadmin' do
+      scenario 'User can create/add barito app' do
+        login_as admin
+
+        barito_app = build(:barito_app)
+        visit app_group_path(@app_group)
+
+        within('#new_barito_app') do
+          fill_in 'barito_app_topic_name', with: barito_app.name
+          fill_in 'barito_app_name', with: barito_app.name
+          fill_in 'barito_app_max_tps', with: 50
+        end
+
+        click_button 'Create'
+        expect(page).to have_content(barito_app.name)
+      end
+    end
+
+    context 'As Authorized User based on Role' do
+      scenario 'User with admin/owner role can create barito app' do
+        create(:app_group_user, app_group: @app_group, role: create(:app_group_role, :admin), user: user_b)
+
+        login_as user_b
+
+        barito_app = build(:barito_app)
+        visit app_group_path(@app_group)
+
+        within('#new_barito_app') do
+          fill_in 'barito_app_topic_name', with: barito_app.name
+          fill_in 'barito_app_name', with: barito_app.name
+          fill_in 'barito_app_max_tps', with: 50
+        end
+
+        click_button 'Create'
+
+        expect(page).to have_content(barito_app.name)
+      end
+
+      scenario 'User with member role cannot create barito app' do
+        create(:app_group_user, app_group: @app_group, role: create(:app_group_role), user: user_b)
+
+        login_as user_b
+
+        barito_app = build(:barito_app)
+        visit app_group_path(@app_group)
+
+        expect(page).not_to have_css('form#new_barito_app input[type="submit"]')
+      end
+    end
+  end
+end
